Move function to own file
